rename tester to testg2c; got rid of main() in favor of testg2c()
